02-Advanced-data-pipelines/03-DBT-Airflow/01-Track-long/pyproject.toml
[tool.poetry]
name = "02-Advanced-data-pipelines/03-DBT-Airflow/01-Advanced-Airflow"
version = "0.1.0"
description = "Data engineering bootcamp by Le Wagon"
authors = ["S. Saunier <seb@lewagon.org>"]

[tool.poetry.dependencies]
python = "^3.8.12"
apache-airflow = "2.3.0"
psycopg2-binary = "^2.9.3"
pandas = "^1.4.2"
pyarrow = "^8.0.0"
apache-airflow-providers-postgres = "4.1.0"
dbt-core = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-mock = "3.2.0"
pytest-responses = "0.4.0"
pendulum = "^2.1.2"
testfixtures = "^6.18.5"
apache-airflow-providers-sqlite = "^2.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
